test: improve tmp file management
