+++ tests/conftest.py
@@ -1,12 +1,17 @@
 """The configuration of the pytest run."""
 
+import json
+import os
 import uuid
 from itertools import product
 from pathlib import Path
+from tempfile import NamedTemporaryFile, TemporaryDirectory
 from typing import Optional
+from urllib.request import urlretrieve
 
 import ee
 import geopandas as gpd
+import pandas as pd
 import pytest
 from shapely import geometry as sg
 
@@ -39,10 +44,11 @@ def tmp_dir() -> Path:
     Returns:
         path to to tmp created directory
     """
-    tmp_dir = Path.home() / "tmp" / "sepal_ui_tests"
-    tmp_dir.mkdir(exist_ok=True, parents=True)
+    with TemporaryDirectory() as tmp_dir:
 
-    return tmp_dir
+        yield Path(tmp_dir)
+
+    return
 
 
 @pytest.fixture(scope="session")
@@ -150,3 +156,193 @@ def alert() -> sw.Alert:
         an alert object
     """
     return sw.Alert()
+
+
+@pytest.fixture(scope="session")
+def fake_vector() -> Path:
+    """Create a fake vector file from the GADM definition of vatican city and save it in the tmp dir.
+
+    Returns:
+        the path to the tmp vector file
+    """
+    with TemporaryDirectory() as tmp_dir:
+        tmp_dir = Path(tmp_dir)
+        link = "https://geodata.ucdavis.edu/gadm/gadm4.1/json/gadm41_VAT_0.json"
+        file = tmp_dir / "gadm41_VAT_0.shp"
+        gpd.read_file(link).to_file(file)
+
+        yield file
+
+    return
+
+
+@pytest.fixture(scope="session")
+def fake_asset(gee_dir: Path) -> Path:
+    """Return the path to a fake asset.
+
+    Returns:
+        the path to the dir
+    """
+    return gee_dir / "feature_collection"
+
+
+@pytest.fixture(scope="session")
+def fake_points() -> Path:
+    """Create a fake point file the tmp file.
+
+    Returns:
+        the path to the point file
+    """
+    with NamedTemporaryFile(suffix=".csv") as tmp_file:
+        tmp_file = Path(tmp_file.name)
+        data = "lat,lon,id\n1,1,0\n0,0,1"
+        tmp_file.write_text(data)
+
+        yield tmp_file
+
+    return
+
+
+@pytest.fixture(scope="session")
+def fake_table() -> Path:
+    """Create a fake table.
+
+    Returns:
+        the path to the created file
+    """
+    with NamedTemporaryFile(suffix=".csv") as tmp_file:
+        tmp_file = Path(tmp_file.name)
+
+        coloseo = [1, 41.89042582290999, 12.492241627092199]
+        fao = [2, 41.88369224629387, 12.489216069409004]
+        columns = ["id", "lat", "lng"]
+        df = pd.DataFrame([coloseo, fao], columns=columns)
+
+        df.to_csv(tmp_file, index=False)
+
+        yield tmp_file
+
+    return
+
+
+@pytest.fixture(scope="session")
+def wrong_table(fake_table: Path) -> Path:
+    """Create a wrongly defined table (with 2 columns instead of the minimal 3.
+
+    Args:
+        fake_table: the path to the complete table
+
+    Returns:
+        the Path to the created file
+    """
+    with NamedTemporaryFile(suffix=".csv") as tmp_file:
+        tmp_file = Path(tmp_file.name)
+        df = pd.read_csv(fake_table).drop(["lng"], axis=1)
+        df.to_csv(tmp_file, index=False)
+
+        yield tmp_file
+
+    return
+
+
+@pytest.fixture(scope="session")
+def readme(root_dir: Path) -> Path:
+    """Return the readme file path.
+
+    Returns:
+        the path to the file
+    """
+    return root_dir / "README.rst"
+
+
+@pytest.fixture(scope="session")
+def file_start() -> str:
+    """The start of any link to the sepal platform.
+
+    Args:
+        the value of the sandbox path
+    """
+    return "https://sepal.io/api/sandbox/jupyter/files/"
+
+
+@pytest.fixture(scope="session")
+def gee_user_dir(gee_dir: Path) -> Path:
+    """Return the path to the gee_dir assets without the project elements.
+
+    Args:
+        gee_dir: the path to the session defined GEE directory
+
+    Returns:
+        the path to gee_dir
+    """
+    legacy_project = Path("projects/earthengine-legacy/assets")
+
+    return gee_dir.relative_to(legacy_project)
+
+
+@pytest.fixture(scope="session")
+def planet_key() -> str:
+    """Get the planet key stored in env.
+
+    Returns:
+        the str key
+    """
+    return os.getenv("PLANET_API_KEY")
+
+
+@pytest.fixture(scope="session")
+def cred() -> list:
+    """Get the credentials stored in env.
+
+    Returns:
+        credential as a list: [cred(username, password)]
+    """
+    credentials = json.loads(os.getenv("PLANET_API_CREDENTIALS"))
+
+    return list(credentials.values())
+
+
+@pytest.fixture(scope="session")
+def image_id() -> str:
+    """The image id of an asset.
+
+    Returns:
+        the AssetId of Daniel Wiell asset
+    """
+    # testing asset from Daniel Wiell
+    # may not live forever
+    return "users/wiell/forum/visualization_example"
+
+
+@pytest.fixture(scope="session")
+def rgb() -> Path:
+    """Add a raster file of the bahamas coming from rasterio test suit.
+
+    Returns:
+        the path to the image
+    """
+    with NamedTemporaryFile(suffix=".tif") as file:
+        file = Path(file.name)
+        link = "https://raw.githubusercontent.com/rasterio/rasterio/master/tests/data/RGB.byte.tif"
+        urlretrieve(link, file)
+
+        yield file
+
+    return
+
+
+@pytest.fixture(scope="session")
+def byte() -> Path:
+    """Add a raster file of the bahamas coming from rasterio test suit.
+
+    Returns:
+        the path to the byte file
+    """
+    with NamedTemporaryFile(suffix=".tif") as file:
+        file = Path(file)
+        link = "https://raw.githubusercontent.com/rasterio/rasterio/master/tests/data/byte.tif"
+        urlretrieve(link, file)
+
+        yield file
+
+    return
